code cleanup and commenting brought about when documenting for jhu report
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@886 1f5c12ca-751b-0410-a591-d2e778427230

+++ moses/src/GenerationDictionary.cpp
@@ -120,7 +120,7 @@ GenerationDictionary::~GenerationDictionary()
 	}
 }
 
-unsigned int GenerationDictionary::GetNumScoreComponents() const
+size_t GenerationDictionary::GetNumScoreComponents() const
 {
   return this->GetNoScoreComponents();
 }
